Dynamically allocate port
I have imported the "net" package to use the net.Listen function.
Instead of using a static port (":8080"), I have changed it to use any available port by passing ":0" to net.Listen.
net.Listen returns a Listener object, which can be used to get the actual port number the server is listening on.
I have updated the log message to display the port number being used by the server.
The http.Serve function is used to serve the HTTP requests using the provided listener, which allows us to serve the content on the dynamically chosen port.