Implement the /api/number endpoint in index.html
Notes from GPT 4:
I have added a button that, when clicked, will make a GET request to /api/number and display the random number in the body of the page. I've also added JavaScript code to handle the button click event, make the request, and update the page background color based on the number's divisibility by 2. If the number is divisible by 2, the background color will be green; otherwise, it will be white.

Resolves #2